+++ clowder/clowder/cli/status_controller.py
@@ -13,7 +13,7 @@ import clowder.util.formatting as fmt
 from clowder.clowder_repo import CLOWDER_REPO
 from clowder.cli.globals import CLOWDER_CONTROLLER
 from clowder.cli.util import run_group_command
-from clowder.util.decorators import network_connection_required
+from clowder.util.connectivity import network_connection_required
 
 
 class StatusController(ArgparseController):
